Move init logic to hook within SetupMessaging component
